(#46) Переписал WeatherService на TypeScript
